Move worklist no owner button to the list title section
