+++ src/js/apps/patients/worklist/filters_app.js
@@ -6,7 +6,7 @@ import App from 'js/base/app';
 
 import AllFiltersApp from 'js/apps/patients/sidebar/filters-sidebar_app';
 
-import { FiltersView, GroupsDropList, NoOwnerToggleView, AllFiltersButtonView } from 'js/views/patients/worklist/filters_views';
+import { FiltersView, GroupsDropList, AllFiltersButtonView } from 'js/views/patients/worklist/filters_views';
 
 const i18n = intl.patients.worklist.filtersApp;
 
@@ -17,10 +17,9 @@ export default App.extend({
   stateEvents: {
     'change:groupId': 'showGroupsFilterView',
   },
-  onStart({ shouldShowOwnerToggle }) {
+  onStart() {
     const currentClinician = Radio.request('bootstrap', 'currentUser');
     this.canViewAssignedActions = currentClinician.can('app:worklist:clinician_filter');
-    this.shouldShowOwnerToggle = shouldShowOwnerToggle;
     this.groups = currentClinician.getGroups();
 
     this.showView(new FiltersView());
@@ -29,7 +28,6 @@ export default App.extend({
   showFilters() {
     this.showAllFiltersButtonView();
     this.showGroupsFilterView();
-    this.showOwnerToggleView();
   },
   showAllFiltersButtonView() {
     if (this.groups.length < 2) return;
@@ -77,17 +75,6 @@ export default App.extend({
     });
 
     this.showChildView('group', groupsFilter);
-  },
-  showOwnerToggleView() {
-    if (!this.shouldShowOwnerToggle || !this.canViewAssignedActions) return;
-
-    const ownerView = this.showChildView('ownerToggle', new NoOwnerToggleView({
-      model: this.getState(),
-    }));
-
-    this.listenTo(ownerView, 'click', () => {
-      this.toggleState('noOwner');
-    });
   },
   _getGroups() {
     const groups = this.groups.clone();
