+++ src/js/views/shared/components/list-search/list-search-component.scss
@@ -1,5 +1,6 @@
 .list-search__container {
   position: relative;
+  width: 166px;
 
   &.is-disabled {
     .icon {
